GPII-2666: Refined "traceback" approach and tests.
