+++ tests/node-tests.js
@@ -3,16 +3,15 @@
 
 require("./js/tests-arraysEqual");
 require("./js/tests-calculateTightness");
-require("./js/tests-collapseSequencesAndCurrentRow");
 require("./js/tests-compare");
 require("./js/tests-compareArrays");
 require("./js/tests-compareMarkdown");
 require("./js/tests-compareObjects");
 require("./js/tests-compareStrings");
-require("./js/tests-compareStringsByLine");
 require("./js/tests-equals");
+require("./js/tests-extractSegments");
 require("./js/tests-longestCommonSequence");
-require("./js/tests-longestDistinctSequenceIds");
+require("./js/tests-longestCommonSequences");
 require("./js/tests-objectsEqual");
 require("./js/tests-singleValueDiff");
 require("./js/tests-sort");
