+++ tests/static/tests-compare.html
@@ -45,7 +45,6 @@
 <script type="text/javascript" src="/tests/js/testDefs-compareArrays.js"></script>
 <script type="text/javascript" src="/tests/js/testDefs-compareObjects.js"></script>
 <script type="text/javascript" src="/tests/js/testDefs-compareStrings.js"></script>
-<script type="text/javascript" src="/tests/js/testDefs-compareStringsByLine.js"></script>
 <script type="text/javascript" src="/tests/js/tests-compare.js"></script>
 </body>
 </html>
