+++ tests/static/tests-extractSegments.html
@@ -1,6 +1,6 @@
 <html>
 <head>
-    <title>"Longest Distinct Sequence IDs" Tests</title>
+    <title>"Extract Segments" Function Tests</title>
 
     <!-- Bring in jQuery and other dependencies -->
     <script type="text/javascript" src="/node_modules/infusion/src/lib/jquery/core/js/jquery.js"></script>
@@ -33,7 +33,7 @@
     <script type="text/javascript" src="/src/js/diff.js"></script>
 </head>
 <body>
-<h1 id="qunit-header">"Longest Distinct Sequence IDs" Tests</h1>
+<h1 id="qunit-header">"Extract Segments" Function Tests</h1>
 <h2 id="qunit-banner"></h2>
 <div id="qunit-testrunner-toolbar"></div>
 <h2 id="qunit-userAgent"></h2>
@@ -41,6 +41,6 @@
 
 <div id="qunit-fixture"></div>
 
-<script type="text/javascript" src="/tests/js/tests-longestDistinctSequenceIds.js"></script>
+<script type="text/javascript" src="/tests/js/tests-extractSegments.js"></script>
 </body>
 </html>
