+++ tests/static/tests-longestCommonSequences.html
@@ -1,6 +1,6 @@
 <html>
 <head>
-    <title>"By Line" String Comparison Tests</title>
+    <title>"Longest Common Sequences" Tests</title>
 
     <!-- Bring in jQuery and other dependencies -->
     <script type="text/javascript" src="/node_modules/infusion/src/lib/jquery/core/js/jquery.js"></script>
@@ -33,7 +33,7 @@
     <script type="text/javascript" src="/src/js/diff.js"></script>
 </head>
 <body>
-<h1 id="qunit-header">"By Line" String Comparison Tests</h1>
+<h1 id="qunit-header">"Longest Common Sequences" Tests</h1>
 <h2 id="qunit-banner"></h2>
 <div id="qunit-testrunner-toolbar"></div>
 <h2 id="qunit-userAgent"></h2>
@@ -41,7 +41,6 @@
 
 <div id="qunit-fixture"></div>
 
-<script type="text/javascript" src="/tests/js/testDefs-compareStringsByLine.js"></script>
-<script type="text/javascript" src="/tests/js/tests-compareStringsByLine.js"></script>
+<script type="text/javascript" src="/tests/js/tests-longestCommonSequences.js"></script>
 </body>
 </html>
