+++ tests/testem.js
@@ -20,8 +20,9 @@ var testemComponent = gpii.testem({
         "tests/static/tests-compareObjects.html",
         "tests/static/tests-compareStrings.html",
         "tests/static/tests-equals.html",
+        "tests/static/tests-extractSegments.html",
         "tests/static/tests-longestCommonSequence.html",
-        "tests/static/tests-longestDistinctSequenceIds.html",
+        "tests/static/tests-longestCommonSequences.html",
         "tests/static/tests-objectsEqual.html",
         "tests/static/tests-singleValueDiff.html",
         "tests/static/tests-sort.html",
